requirements.txt
watchdog
virustotal-api
requests
elasticsearch
python-magic
